Update Folder Transaksi

+++ Web/transaksi/detailtransaksi.php
@@ -4,13 +4,12 @@ require ("../koneksi.php");
 
 session_start();
 
-if(!isset($_SESSION['id'])){
-    $_SESSION['msg'] = 'Anda harus login untuk mengakses halaman ini!';
-    header('Location: login.php');
-}
+//session
 $sesID = $_SESSION['id'];
 $sesName = $_SESSION['username'];
 $sesLvl = $_SESSION['role'];
+$sesImg = $_SESSION['foto'];
+$path = '../images/admin/';
 
 ?>
 <!DOCTYPE html>
@@ -20,9 +19,9 @@ $sesLvl = $_SESSION['role'];
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=Edge">
     <meta content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no" name="viewport">
-    <title>Order | SUGAR CANE</title>
+    <title>Transaction Details | SUGAR CANE</title>
     <!-- Favicon-->
-    <link rel="icon" href="favicon.ico" type="image/x-icon">
+    <link rel="icon" href="../favicon.ico" type="image/x-icon">
 
     <!-- Google Fonts -->
     <link href="https://fonts.googleapis.com/css?family=Roboto:400,700&subset=latin,cyrillic-ext" rel="stylesheet" type="text/css">
@@ -36,9 +35,6 @@ $sesLvl = $_SESSION['role'];
 
     <!-- Animation Css -->
     <link href="../plugins/animate-css/animate.css" rel="stylesheet" />
-
-    <!-- Morris Chart Css-->
-    <link href="../plugins/morrisjs/morris.css" rel="stylesheet" />
 
     <!-- Custom Css -->
     <link href="../css/style.css" rel="stylesheet">
@@ -77,14 +73,13 @@ $sesLvl = $_SESSION['role'];
                    <!-- User Info -->
                     <li class="dropdown">
                         <a href="javascript:void(0);" class="dropdown-toggle" data-toggle="dropdown" role="button">
-                            <div class="name" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"></div>
-                            <img class="img-profile rounded-circle" src="../images/user.png" width="70%" style="border-radius: 50px;">
+                            <img class="img-profile rounded-circle" src="<?php echo $path.$sesImg; ?>" width="36" height="36" style="border-radius: 50px; margin-top: -5px; margin-left: 5px;" >
                         </a>
 
                         <!-- Dropdown - User Information -->
                         <ul class="dropdown-menu" style="border-radius: 5px;">
                             <div class="dropdown-divider"></div>
-                            <li><a href="editprofile.php"><i class="material-icons">person</i>Profile</a></li> 
+                            <li><a href="../editprofile.php"><i class="material-icons">person</i>Profile</a></li> 
                         </ul>
                     </li>
                     <!-- #User Info -->
@@ -104,35 +99,35 @@ $sesLvl = $_SESSION['role'];
                         <li>
                             <a href="../index.php">
                                 <i class="material-icons">home</i>
-                                <span>Dashboard</span>
+                                <span>DASHBOARD</span>
                             </a>
                         </li>
                         <li>
                             <a href="../admin/adminhome.php">
                                 <i class="material-icons">account_box</i>
-                                <span>Admins</span>
+                                <span>ADMIN</span>
                             </a>
                         </li>
                         <li>
                             <a href="../user/userhome.php">
                                 <i class="material-icons">person</i>
-                                <span>Users</span>
+                                <span>USER</span>
                             </a>
                         </li>
                         <li>
                             <a href="javascript:void(0);" class="menu-toggle">
                                 <i class="material-icons">library_books</i>
-                                <span>Data Barang</span>
+                                <span>PRODUCT</span>
                             </a>
                             <ul class="ml-menu">
                                 <li>
                                     <a href="../barang/baranghome.php">
-                                        <span>Barang</span>
+                                        <span>PRODUCT</span>
                                     </a>
                                 </li>
                                 <li>
                                     <a href="../barang/detailukuran.php">
-                                        <span>Detail Ukuran</span>
+                                        <span>SIZE DETAILS</span>
                                     </a>
                                 </li>
                             </ul>
@@ -140,17 +135,17 @@ $sesLvl = $_SESSION['role'];
                         <li class="active">
                             <a href="javascript:void(0);" class="menu-toggle">
                                 <i class="material-icons">assessment</i>
-                                <span>Transaksi</span>
+                                <span>TRANSACTION</span>
                             </a>
                             <ul class="ml-menu">
                                 <li>
                                     <a href="transaksihome.php">
-                                        <span>Transaksi</span>
+                                        <span>TRANSACTION</span>
                                     </a>
                                 </li>
                                 <li class="active">
                                     <a href="detailtransaksi.php">
-                                        <span>Order</span>
+                                        <span>TRANSACTION DETAILS</span>
                                     </a>
                                 </li>
                             </ul>
@@ -175,7 +170,7 @@ $sesLvl = $_SESSION['role'];
                     <div class="card">
                         <div class="header">
                             <h2>
-                                ORDER
+                                TRANSACTION DETAILS
                             </h2>
                             <a href="dtransaksicreate.php"> 
                                 <button type="button" class="btn bg-light-green waves-effect" style="border-radius: 3px;">
@@ -193,15 +188,34 @@ $sesLvl = $_SESSION['role'];
                                             <th>ID Pesanan</th>
                                             <th>ID Transaksi</th>
                                             <th>ID Barang</th>
+                                            <th>Varian</th>
+                                            <th>ID Detail Ukuran</th>
+                                            <th>Harga</th>
                                             <th>Quantity</th>
                                             <th>Sub Harga</th>
                                             <th>Menu</th>
                                         </tr>
                                     </thead>
                                     <?php 
-                                        $query = "SELECT * FROM pesanan";
+                                        $batas = 10;
+                                        $halaman = isset($_GET['halaman'])?(int)$_GET['halaman'] : 1;
+                                        $halaman_awal = ($halaman>1) ? ($halaman * $batas) - $batas : 0;    
+
+                                        $previous = $halaman - 1;
+                                        $next = $halaman + 1;
+                                        
+                                        $data = mysqli_query($koneksi, "SELECT * FROM pesanan");
+                                        $jumlah_data = mysqli_num_rows($data);
+                                        $total_halaman = ceil($jumlah_data / $batas);
+
+                                        $query = "SELECT pesanan.id_pesanan, pesanan.id_transaksi, pesanan.id_barang, transaksi.id_transaksi, barang.id_barang, barang.varian, barang.id_detailukuran, detailukuran.id_detailukuran, detailukuran.harga, pesanan.qty, pesanan.subharga
+                                            FROM pesanan
+                                            INNER JOIN transaksi ON pesanan.id_transaksi = transaksi.id_transaksi
+                                            INNER JOIN barang ON pesanan.id_barang = barang.id_barang
+                                            INNER JOIN detailukuran ON barang.id_detailukuran = detailukuran.id_detailukuran
+                                            limit $halaman_awal, $batas";
                                         $result = mysqli_query($koneksi, $query);
-                                        $no = 1;
+                                        $no = $halaman_awal+1;
 
                                         if ($sesLvl == 1){
                                             $dis = "";
@@ -209,36 +223,60 @@ $sesLvl = $_SESSION['role'];
                                             $dis = "disabled";
                                         }
 
+                                        //Menampilkan data dalam database menggunakan array
                                         while ($row = mysqli_fetch_array($result)){
-                                            $id_pesanan = $row['id_pesanan'];
-                                            $id_transaksi = $row['id_transaksi'];
-                                            $id_barang= $row['id_barang'];
-                                            $qty= $row['qty'];
-                                            $subharga= $row['subharga'];
+
                                     ?>
                                     <tbody>
                                         <tr>
-                                            <td><?php echo $no; ?></td>
-                                            <td><?php echo $id_pesanan; ?></td>
-                                            <td><?php echo $id_transaksi; ?></td>
-                                            <td><?php echo $id_barang; ?></td>
-                                            <td><?php echo $qty; ?></td>
-                                            <td><?php echo $subharga; ?></td>
+                                            <td><?php echo $no++; ?></td>
+                                            <td><?php echo $row['id_pesanan']; ?></td>
+                                            <td><?php echo $row['id_transaksi']; ?></td>
+                                            <td><?php echo $row['id_barang']; ?></td>
+                                            <td><?php echo $row['varian']; ?></td>
+                                            <td><?php echo $row['id_detailukuran']; ?></td>
+                                            <td><?php echo $row['harga']; ?></td>
+                                            <td><?php echo $row['qty']; ?></td>
+                                            <td><?php echo $row['subharga']; ?></td>
                                             <td>
                                                 <a href="dtransaksiedit.php?id=<?php echo $row['id_pesanan']; ?>">
                                                     <input type="button" class="btn btn-info" value="Edit" name="edit" <?php echo $dis; ?>>
                                                 </a>
-                                                <a href="dtransaksidelete.php?id=<?php echo $row['id_pesanan']; ?>">
+                                                <a href="dtransaksidelete.php?id=<?php echo $row['id_pesanan'];?>" onclick="return confirm('Do you want to delete these records? This action cannot be undone. You will be unable to recover any data.');">
                                                     <input type="button" class="btn btn-danger" value="Delete" name="delete" <?php echo $dis; ?>>
                                                 </a>
                                             </td>
                                         </tr>
                                         <?php 
-                                            $no++;
-                                        }
+                                            }
                                         ?>
                                     </tbody>
                                 </table>
+                                <nav>
+                                <ul class="pagination table-bordered">
+                                    <li>
+                                        <a class="waves-effect" <?php if($halaman > 1){ echo "href='detailtransaksi.php?halaman=$previous'";} ?> >
+                                            PREVIOUS
+                                            <!-- <i class="material-icons">chevron_left</i> -->
+                                        </a>
+                                    </li>
+
+                                    <?php 
+                                        for($x=1; $x<=$total_halaman; $x++){
+                                    ?>
+                                        <li><a class="waves-effect" href="detailtransaksi.php?halaman=<?php echo $x ?>"><?php echo $x; ?><!-- &nbsp;&nbsp;| --></a></li>
+                                    <?php
+                                        }
+                                    ?>  
+
+                                    <li>
+                                        <a class="waves-effect" <?php if($halaman < $total_halaman){ echo "href='detailtransaksi.php?halaman=$next'";} ?>>
+                                            NEXT
+                                            <!-- <i class="material-icons">chevron_right</i> -->
+                                        </a>
+                                    </li>
+                                </ul>
+                                </nav>
                             </div>
                         </div>
                     </div>
@@ -253,7 +291,7 @@ $sesLvl = $_SESSION['role'];
                 <!-- konten modal-->
                     <div class="modal-content">
                         <!-- heading modal -->
-                        <div class="modal-header">
+                        <div class="modal-header" style="background: #FFCCCC;">
                             <h3 class="modal-title" id="modallogoutLabel">Confirm Logout</h3
                                 >
                         </div>
@@ -264,8 +302,8 @@ $sesLvl = $_SESSION['role'];
                         <!-- footer modal -->
                         <div class="modal-footer">
                             <a href="../logout.php">
-                                <button type="button" class="btn btn-link waves-effect">Yes</button>
-                                <button type="button" class="btn btn-link waves-effect" data-dismiss="modal">Cancel</button>
+                                <button type="button" class="btn btn-danger waves-effect">Yes</button>
+                                <button type="button" class="btn btn-primary waves-effect" data-dismiss="modal">Cancel</button>
                             </a>
                         </div>
                     </div>
@@ -294,10 +332,8 @@ $sesLvl = $_SESSION['role'];
     <script src="../plugins/jquery-datatable/extensions/export/dataTables.buttons.min.js"></script>
     <script src="../plugins/jquery-datatable/extensions/export/buttons.flash.min.js"></script>
     <script src="../plugins/jquery-datatable/extensions/export/jszip.min.js"></script>
-    <script src="../plugins/jquery-datatable/extensions/export/pdfmake.min.js"></script>
     <script src="../plugins/jquery-datatable/extensions/export/vfs_fonts.js"></script>
     <script src="../plugins/jquery-datatable/extensions/export/buttons.html5.min.js"></script>
-    <script src="../plugins/jquery-datatable/extensions/export/buttons.print.min.js"></script>
 
     <!-- Custom Js -->
     <script src="../js/admin.js"></script>
